+++ css/profile.css
@@ -28,6 +28,11 @@
   background-color: var(--dark-primary);
 }
 
+#profile-avatar-status {
+  width: 100%;
+  height: 100%;
+}
+
 /* 
   activity
 */
