+++ html/gfx.html
@@ -1,12 +1,12 @@
 <h1>Graphics Design</h1>
-<a class="card hover" href="https://discord.gg/gYAW7JF" title="Quantum Society Discord server" target="_blank"
+<a class="card hover" href="https://discord.gg/gYAW7JF" title="Quantum Society Discord Server" target="_blank"
   rel="noopener noreferrer" style="background-color: rgba(29, 91, 138, 0.25)">
-  <img alt="Quantum Society Icon" class="card-icon circle" src="assets/projects/quantum-society.gif" />
+  <img alt="Quantum Society Discord server icon" class="card-icon circle" src="assets/projects/quantum-society.gif" />
   <div class="card-content">
-    <h3>Quantum Society</h3>
+    <h3>Quantum Society Discord Server</h3>
     <p>
-      My Discord server, all art in it including the penguin emotes (<img alt="Penguin Emoji" class="emoji"
-        src="assets/emojis/penguin.webp" />
+      My personal Discord server. I created all the graphics including the penguin emotes. (<img alt="Penguin Emoji"
+        class="emoji" src="assets/emojis/penguin.webp" />
       <img alt="Penguin Heart Emoji" class="emoji" src="assets/emojis/penguin-heart.png" />
       <img alt="Penguin Spin Emoji" class="emoji" src="assets/emojis/penguin-spin.gif" />
       <img alt="Penguin Party Emoji" class="emoji" src="assets/emojis/penguin-party.gif" />
@@ -14,19 +14,35 @@
     </p>
   </div>
 </a>
-<a class="card hover" href="https://discord.gg/ps" title="PlayStation Discord server" target="_blank"
+<a class="card hover" href="https://discord.gg/ps" title="PlayStation Discord Server" target="_blank"
   rel="noopener noreferrer" style="background-color: rgba(8, 15, 138, 0.25)">
-  <img alt="PlayStation Discord Icon" class="card-icon circle" src="assets/projects/playstation.gif" />
+  <img alt="PlayStation Discord server icon" class="card-icon circle" src="assets/projects/playstation.gif" />
   <div class="card-content">
-    <h3>PlayStation Discord</h3>
-    <p>The icon for the community run PlayStation Discord server</p>
+    <h3>PlayStation Discord Server</h3>
+    <p>The community run PlayStation Discord server. I created the icon for this community.</p>
   </div>
 </a>
-<a class="card hover" href="https://discord.gg/ian" title="Sol Sanctum Discord server" target="_blank"
+<a class="card hover" href="https://discord.gg/ian" title="Sol Sanctum Discord Server" target="_blank"
   rel="noopener noreferrer" style="background-color: rgba(171, 55, 206, 0.25)">
-  <img alt="Sol Sanctum Discord Icon" class="card-icon circle" src="assets/projects/ian.gif" />
+  <img alt="Sol Sanctum Discord server icon" class="card-icon circle" src="assets/projects/ian.gif" />
   <div class="card-content">
-    <h3>Sol Sanctum Discord</h3>
-    <p>The icon and some stickers for the Sol Sanctum Discord server</p>
+    <h3>Sol Sanctum Discord Server</h3>
+    <p>I created the icon and a few stickers for this community.</p>
+  </div>
+</a>
+<a class="card hover" href="https://discord.gg/bird" title="Birb Paradise Discord Server" target="_blank"
+  rel="noopener noreferrer" style="background-color: rgba(60, 206, 55, 0.25)">
+  <img alt="Birb Paradise Discord server icon" class="card-icon circle" src="assets/projects/birb.gif" />
+  <div class="card-content">
+    <h3>Birb Paradise Discord Server</h3>
+    <p>I created the icon for this community.</p>
+  </div>
+</a>
+<a class="card hover" href="https://www.reddit.com/r/keycapdesigners" title="Keycap Designers Subreddit" target="_blank"
+  rel="noopener noreferrer" style="background-color: rgba(206, 55, 55, 0.25)">
+  <img alt="r/keycapdesigners icon" class="card-icon circle" src="assets/projects/keycap-designers.gif" />
+  <div class="card-content">
+    <h3>Keycap Designers Subreddit</h3>
+    <p>I created the animated version of the icon for this community.</p>
   </div>
 </a>
