+++ js/index.ts
@@ -2,8 +2,8 @@ import "./events/keydown";
 import "./events/touch";
 import "./events/wheel";
 
-import "./extra/blog";
-import "./extra/contact";
-import "./extra/guide";
-import "./extra/router";
-import "./extra/ws";
+import "./modules/blog";
+import "./modules/contact";
+import "./modules/guide";
+import "./modules/router";
+import "./modules/presence";
